requirements.txt
flask
python-dotenv
cohere
scikit-learn
gunicorn
onnxruntime